Create ComputeChange.java
Write a simple java program to be used by a shopkeeper in Kisii. Once a customer visits a shop
& Identifies items to buy, the attendant totals up the price of the goods and informs the customer
cost of all goods. The customer tenders in the amount of money which is normally more that the
cost of the goods. Once the attendant receives the tendered amount, he/she calculates the change
and then hands over the goods plus the change to the customer.
You are required to write a java class (ComputeChange.java) with the main method which takes
in the total cost of goods, and the amount tendered by the customer. The application should
compute and return the change to be given to the customer. The application should also print out
a receipt for the customer.
Activity 2
Modify the class in Activity 1 such that the shop attendant also enters the details and quantities
of Items bought and a comprehensive receipt printed out. I.e. when the class runs, the following
should be the output. Note: Highlighted items are those that a user should be prompted to enter.
==================================
Welcome to Musa buddy shop - Kisii
===================================
Please enter the number of items bought by customer: 3
******************* Item One ************
Enter description of Item 1: Sugar
Enter Quantity of Item 1: 4
Enter unit cost of item 1 : 250
******************** Item two ************
Enter description of item 2: Soap
Enter quantity of Item 2: 5
Enter unit cost of item 2 : 150
********************** Item three ********
Enter description of item 3 : Cups
Enter quantity of Item 3 : 2
Enter unit cost of item 3 : 300
*********************************************
---------------------------------------------
Summary of items bought
---------------------------------------------
Quantity Description UnitCost SubTotal
---------------------------------------------
4 Sugar 250 1000
5 Soap 150 750
2 Cups 300 600
---------------------------------------------
The total Cost of items is : 2350
Please enter the amount tendered by customer : 3000
The change due to customer is : 650
Printing Customer receipt!!!!!!!!!!
=============================================
=========== Musa buddy shop – Kikuubo========
=============================================
Date: 20 – 03 -2021 Time: 1400 hr
Quantity Description UnitCost SubTotal
---------------------------------------------
4 Sugar 250 1000
5 Soap 150 750
2 Cups 300 600
------------------------------------------------
Total cost of goods 2350 +
Balance to customer 650 -
Amount tendered by customer 3000
------------------------------------------------
Thank you ! Come again
Goods once sold are not returnable